refactor(agents): extract imageResult helpers to server-only image-result.ts
