+++ src/agents/tools/common.ts
@@ -1,8 +1,4 @@
-import fs from "node:fs/promises";
 import type { AgentTool, AgentToolResult } from "@mariozechner/pi-agent-core";
-import { detectMime } from "../../media/mime.js";
-import type { ImageSanitizationLimits } from "../image-sanitization.js";
-import { sanitizeToolResultImages } from "../tool-images.js";
 
 // oxlint-disable-next-line typescript/no-explicit-any
 export type AnyAgentTool = AgentTool<any, unknown> & {
@@ -225,51 +221,4 @@ export function wrapOwnerOnlyToolExecution(
       throw new Error(OWNER_ONLY_TOOL_ERROR);
     },
   };
-}
-
-export async function imageResult(params: {
-  label: string;
-  path: string;
-  base64: string;
-  mimeType: string;
-  extraText?: string;
-  details?: Record<string, unknown>;
-  imageSanitization?: ImageSanitizationLimits;
-}): Promise<AgentToolResult<unknown>> {
-  const content: AgentToolResult<unknown>["content"] = [
-    {
-      type: "text",
-      text: params.extraText ?? `MEDIA:${params.path}`,
-    },
-    {
-      type: "image",
-      data: params.base64,
-      mimeType: params.mimeType,
-    },
-  ];
-  const result: AgentToolResult<unknown> = {
-    content,
-    details: { path: params.path, ...params.details },
-  };
-  return await sanitizeToolResultImages(result, params.label, params.imageSanitization);
-}
-
-export async function imageResultFromFile(params: {
-  label: string;
-  path: string;
-  extraText?: string;
-  details?: Record<string, unknown>;
-  imageSanitization?: ImageSanitizationLimits;
-}): Promise<AgentToolResult<unknown>> {
-  const buf = await fs.readFile(params.path);
-  const mimeType = (await detectMime({ buffer: buf.slice(0, 256) })) ?? "image/png";
-  return await imageResult({
-    label: params.label,
-    path: params.path,
-    base64: buf.toString("base64"),
-    mimeType,
-    extraText: params.extraText,
-    details: params.details,
-    imageSanitization: params.imageSanitization,
-  });
 }
